# api/openapi.yaml
openapi: 3.0.0
info:
  title: Finsolvz Backend API
  description: |
    Comprehensive financial solutions management system with JWT authentication, 
    user management, company management, report type management, and complete report management.
    
    **Authentication Required**: Most endpoints require a valid JWT token in the Authorization header.
    
    **Roles & Authorization**:
    - `SUPER_ADMIN`: Full system access including user registration, role management, and company/user operations
    - `ADMIN`: User management access (get users list)  
    - `CLIENT`: Basic authenticated access to reports and own profile
    
    **Authorization Pattern**: Authorization checks are performed at the controller level, not middleware level.
    Most report endpoints require authentication only, while user management requires specific roles.
    
    **Base URL**: `http://localhost:8787` (development)
  version: 2.0.0
  contact:
    name: Finsolvz Team
    email: support@finsolvz.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8787
    description: Local Development Server
  - url: https://finsolvz-backend.example.com
    description: Production Environment
...
paths:
...
  /api/register:
    post:
      summary: Register new user (SUPER_ADMIN only)
      operationId: register
      tags:
        - User Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Success"
                  newUser:
                    $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/users:
    get:
      summary: Get all users (ADMIN+ only)
      operationId: getUsers
      tags:
        - User Management
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/users/{id}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags:
        - User Management
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "60f1b2e5e4b0c7a1d8b9c0d1"
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update user (SUPER_ADMIN only)
      operationId: updateUser
      tags:
        - User Management
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "60f1b2e5e4b0c7a1d8b9c0d1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User updated"
                  updatedUser:
                    $ref: '#/components/schemas/UserResponse'

    delete:
      summary: Delete user (SUPER_ADMIN only)
      operationId: deleteUser
      tags:
        - User Management
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "60f1b2e5e4b0c7a1d8b9c0d1"
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Success"
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/loginUser:
    get:
      summary: Get current authenticated user
      operationId: getLoginUser
      tags:
        - User Management
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /api/updateRole:
    put:
      summary: Update user role (SUPER_ADMIN only)
      operationId: updateRole
      tags:
        - User Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Success"
                  user:
                    $ref: '#/components/schemas/UserResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/change-password:
    patch:
      summary: Change current user password
      operationId: changePassword
      tags:
        - User Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password successfully changed"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - name
        - email
        - password
        - role
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        password:
          type: string
          format: password
          minLength: 6
          example: "securePassword123!"
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, CLIENT]
          example: "CLIENT"
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: "John Updated"
        email:
          type: string
          format: email
          example: "john.updated@example.com"
        password:
          type: string
          format: password
          minLength: 6
          example: "newPassword123!"
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, CLIENT]
          example: "ADMIN"
    UpdateRoleRequest:
      type: object
      required:
        - userId
        - newRole
      properties:
        userId:
          type: string
          example: "60f1b2e5e4b0c7a1d8b9c0d1"
        newRole:
          type: string
          enum: [SUPER_ADMIN, ADMIN, CLIENT]
          example: "ADMIN"
    ChangePasswordRequest:
      type: object
      required:
        - newPassword
        - confirmPassword
      properties:
        newPassword:
          type: string
          format: password
          minLength: 6
          example: "newPassword123!"
        confirmPassword:
          type: string
          format: password
          minLength: 6
          example: "newPassword123!"
    UserResponse:
      type: object
      properties:
        _id:
          type: string
          example: "60f1b2e5e4b0c7a1d8b9c0d1"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, CLIENT]
          example: "CLIENT"
        company:
          type: array
          items:
            type: string
          example: ["60f1b2e5e4b0c7a1d8b9c0d2"]
        createdAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"

    # Company Management Schemas
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Application-specific error code
          example: "INVALID_INPUT"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request payload"
        details:
          type: string
          description: Additional error details
          example: "Validation failed: email is required"
  responses:
...
    ForbiddenError:
      description: Access denied due to insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "FORBIDDEN"
            message: "Access denied"
            details: "Insufficient permissions for this operation"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "NOT_FOUND"
            message: "Resource not found"
            details: "The requested resource could not be found"
    ConflictError:
      description: Resource conflict or already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "CONFLICT"
            message: "Resource conflict"
            details: "Email already registered"
...